feat: :art: UI Update and Component Enhancements
This commit includes changes to the project's user interface, which encompass:  Modifications to the design of the main page, adding the Carousel component with its slides. Conversion of Carousel.jsx and SlidesButtons.jsx components to .tsx. Introduction of the new Topbar.tsx component. Addition of the slides.ts file to manage resources related to slides within the application. These changes are intended to enhance the appearance and functionality of the user interface, as well as to keep the project up-to-date with more efficient and relevant components.