feat: :art: Hero Component, Styling, and Refactoring
This commit encompasses various updates and enhancements:  - Created the 'Hero' component. - Added an image for the hero in the Tailwind configuration. - Fixed the navbar at the top of the screen and applied a gradient background, configured in Tailwind. - Changed the primary color to the secondary color, affecting the footer, cards, navbar, and other components that used the primary color as the background. - Added 'mt' (margin-top) to all pages due to the 'position: fixed' applied to the navbar. - Created the 'BookNowButton' component. - Stopped using the 'Topbar' component for design purposes but retained it in the components folder.