feat: :art: Scroll-Responsive Navbar Styling
This commit adds functionality to the navbar:  Initially, the navbar is visible at the top of the screen with a transparent background (bg-transparent). When the user scrolls down, the navbar background changes to primary (bg-primary). If the user continues scrolling, the navbar hides. When the user scrolls up, the navbar reappears with a primary background. When the user scrolls up further and reaches the top of the document, the navbar returns to a transparent background.