fix: :art: Enable 'pages' in NexthAuth configuration and :art: Add 'target' attribute to gallery page links
Enabled the 'pages' setting in NexthAuth configuration to address the issue of not redirecting to the login page in production. Added the 'target' attribute to links on the gallery page.

This commit addresses the following changes:  Fixed the issue of not redirecting to the login page in production by enabling the 'pages' setting in NexthAuth configuration. Added the 'target' attribute to links on the gallery page to specify how the linked pages should be opened.