refactor: :building_construction: Add new AdminPage, add session buttons
This commit introduces several structural and navigational improvements to the project:  Added the 'AdmiPage' to manage customers. Links pointing to '/profile' have been redirected to this new page. The temporary 'profile' page has been removed.  Included 'LoginButton' and 'SignOutButton' to separate the login process from the navbar, improving user interaction and clarity.  Renamed the 'about-us' page to 'salon about' to better reflect its content.

Introduction of 'AdmiPage' for customer management. Redirection of links to '/profile' to the new 'AdmiPage.' Removal of the temporary 'profile' page. Addition of 'LoginButton' and 'SignOutButton' for a cleaner navbar. Renaming the 'about-us' page to 'salon about' for clarity.